fix(merge-guard): share the read floor's danger/compound predicates with the mint (4 pre-existing under-blocks)
An independent blind clean-room re-pass on the minimal mint gate surfaced four
pre-existing under-blocks the gate did not cover. Each was reproduced before
fixing. The mint side now shares the read floor's danger and compound
predicates by construction, so a token can never mint for something the read
floor would not gate (mint is a subset of read).

- Classifier under-block: the mint minted a token for any command the
  operation classifier recognized, even when the read floor would not gate it
  (a bare reversible close, or non-command API prose). The mint now gates its
  token-write on the shared danger predicate. Promotes is_dangerous_command,
  is_compound_destructive_command, the dangerous-command patterns, and the
  strip helpers to shared/merge_guard_common.py; merge_guard_pre.py re-imports
  them, so the read floor is byte-identical to before.

- Close binding bypass: a token approved for a bare reversible close
  authorized the irreversible branch-deleting variant, because the deleting
  flag was bound on merge but not on close. Binds the branch-delete flag on
  close.

- Push operation collapse: a plain push to the default branch and a forced
  push both classified as the same forced-push operation, so a plain-push
  approval could authorize a history-rewriting forced push. Plain push to the
  default branch is now its own gated operation, distinct from forced push;
  operation identity keeps the two from cross-authorizing, and a faithful
  plain-push token authorizes its own execution.

- Compound mint integrity: a chained command in one clicked option minted a
  token for just the recognized leg, letting an arbitrary second command ride
  along. The mint now refuses any clicked option whose located command region
  is a compound (mirroring the read side). The surrounding option prose is not
  scanned, so a faithful single-command option whose description merely
  contains a separator or newline still mints.

Read floor behavior is unchanged (verified by byte-identity of the relocated
predicates); the shared import stays inside the deny-guarded block so a load
failure fails closed; the approval-to-command binding is unchanged, so no
escalation or substitution is admitted; faithful single-command approvals all
mint.
