+++ pact-plugin/tests/test_merge_guard_1037_narrow.py
@@ -132,7 +132,9 @@ class TestNarrowGhCommentCarrierStripBlock:
 
         cmd = 'gh pr comment 5 --body "$(gh pr merge 5)"'
         assert is_dangerous_command(cmd)
-        monkeypatch.setattr("merge_guard_pre._has_command_substitution", lambda q: False)
+        # is_dangerous_command + its _has_command_substitution helper are promoted to
+        # shared (GAP1); neuter the SHARED definition (pre re-exports the same object).
+        monkeypatch.setattr("shared.merge_guard_common._has_command_substitution", lambda q: False)
         assert not is_dangerous_command(cmd)  # flips ALLOW under the neuter
 
     def test_escaped_quote_outside_carrier_flag_blocks(self):
